# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  description: Projeto do Processo Seletivo Halliburton
  title: Boarding Manager API
  contact:
    email: igorfs.pro@gmail.com
    name: Igor Santos
tags:
- name: Employees
  description: API Employees of Oil and Gas Technicians
paths:
  /employees:
    get:
      tags:
      - Employees
      summary: Employees List
      description: <p>List all Employees according to the defined landed period.</p>
      operationId: findAll
      parameters:
      - $ref: '#/components/parameters/periodBeginningQueryParam'
      - $ref: '#/components/parameters/periodEndingQueryParam'
      - $ref: '#/components/parameters/pageQueryParam'
      - $ref: '#/components/parameters/sizeQueryParam'
      responses:
        200:
          description: Employees List.
          headers:
            Link:
              $ref: '#/components/headers/link'
            Total-Elements:
              $ref: '#/components/headers/totalElements'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePage'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        422:
          description: Unprocessable Entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyError'
    post:
      tags:
      - Employees
      summary: Add a new Employee
      operationId: addOne
      requestBody:
        $ref: '#/components/requestBodies/employee'
      responses:
        201:
          description: Employee added successfully
          headers:
            Location:
              $ref: '#/components/headers/location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        422:
          description: Unprocessable Entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyError'
  /employees/{cpf}:
    get:
      tags:
      - Employee
      operationId: findOne
      summary: find an Employee by cpf
      parameters:
      - $ref: '#/components/parameters/cpfEmployeePathParam'
      responses:
        200:
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        404:
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags:
      - Employee
      summary: Update the Employee Boarding Status
      operationId: mergePatch
      parameters:
      - $ref: '#/components/parameters/cpfEmployeePathParam'
      requestBody:
        $ref: '#/components/requestBodies/employeeMergePatch'
      responses:
        200:
          description: Boarding Status Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        422:
          description: Unprocessable Entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyError'
components:
  headers:
    totalElements:
      $ref: 'models/schemas/headers/total-elements.yaml#/Total-Elements'
    link:
      $ref: 'models/schemas/headers/link.yaml#/Link'
    location:
      $ref: 'models/schemas/headers/location.yaml#/Location'
  schemas:
    Employee:
      $ref: models/employee/employee.yaml#/Employee
    EmployeeMergePatch:
      $ref: models/employee/employee-merge-patch.yaml#/EmployeeMergePatch
    EmployeePage:
      $ref: models/employee/employee-page.yaml#/EmployeePage
    Problem:
      $ref: models/schemas/problem-details.yaml#/ProblemDetails
    PropertyError:
      $ref: models/schemas/property-error.yaml#/PropertyError
  parameters:
    cpfEmployeePathParam:
      in: path
      name: cpf
      description: The Employee's CPF
      required: true
      schema:
        type: string
        minLength: 11
        maxLength: 11 
    periodBeginningQueryParam:
      in: query
      name: periodBeginning
      description: period's first day to be verified
      required: false
      schema:
        type: string
        format: full-date
        example: 2021-06-23T00:00:00Z
    periodEndingQueryParam:
      in: query
      name: periodEnding
      description: period's last day to be verified
      required: false
      schema:
        type: string
        format: full-date
        example: 2021-06-23T00:00:00Z
    pageQueryParam:
      $ref: models/schemas/parameters/query/page.yaml#/PageQueryParam
    sizeQueryParam:
      $ref: models/schemas/parameters/query/size.yaml#/SizeQueryParam
  requestBodies:
    employee:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Employee'
    employeeMergePatch:
      required: true
      content:
        application/merge-patch+json:
          schema:
            $ref: '#/components/schemas/EmployeeMergePatch'